requirements.txt
yfinance
pandas
numpy
textblob
colorama
tabulate
google-generativeai>=0.3.0
jinja2
